move scripts and styles
