+++ blocks/form-handler/controllers/event-info-component-controller.js
@@ -1,5 +1,5 @@
 /* eslint-disable no-use-before-define */
-import { getLibs } from '../../../scripts/utils.js';
+import { getLibs } from '../../../events/scripts/utils.js';
 import { changeInputValue } from '../../../utils/utils.js';
 
 const { createTag } = await import(`${getLibs()}/utils/utils.js`);
