+++ components/partner-selector-group/partner-selector-group.js
@@ -1,4 +1,4 @@
-import { getLibs } from '../../scripts/utils.js';
+import { getLibs } from '../../events/scripts/utils.js';
 import { isEmptyObject } from '../../utils/utils.js';
 import { style } from './partner-selector-group.css.js';
