+++ components/partner-selector/partner-selector.js
@@ -1,4 +1,4 @@
-import { getLibs } from '../../scripts/utils.js';
+import { getLibs } from '../../events/scripts/utils.js';
 import { style } from './partner-selector.css.js';
 import { createPartner } from '../../utils/esp-controller.js';
