+++ components/product-selector/product-selector.css.js
@@ -1,5 +1,5 @@
 /* stylelint-disable selector-class-pattern */
-import { getLibs } from '../../scripts/utils.js';
+import { getLibs } from '../../events/scripts/utils.js';
 
 const { css } = await import(`${getLibs()}/deps/lit-all.min.js`);
