+++ components/product-selector/product-selector.js
@@ -1,4 +1,4 @@
-import { getLibs } from '../../scripts/utils.js';
+import { getLibs } from '../../events/scripts/utils.js';
 import { isEmptyObject } from '../../utils/utils.js';
 import { style } from './product-selector.css.js';
