+++ events/scripts/scripts.js
@@ -11,7 +11,7 @@
  */
 
 import { setLibs, decorateArea } from './utils.js';
-import { lazyCaptureProfile } from '../utils/event-apis.js';
+import { lazyCaptureProfile } from '../../utils/event-apis.js';
 
 const locales = {
   '': { ietf: 'en-US', tk: 'jdq5hay.css' },
@@ -65,7 +65,7 @@ decorateArea();
 
 const miloLibs = setLibs(LIBS);
 
-window.bm8r = await import('../deps/block-mediator.min.js').then((mod) => mod.default);
+window.bm8r = await import('../../deps/block-mediator.min.js').then((mod) => mod.default);
 
 (function loadStyles() {
   const paths = [`${miloLibs}/styles/styles.css`];
